FUEL@OPNFV CI EVOLUTION
A very early sample approach to decouple Fuel@OPNFV core builds from
plug-in builds. For the Goal objectives - see JIRA: FUEL-207.
Note that this is an experimental work together with the releng project.

This is an early peace of work - not verified e2e, but rather intended
to show intents.
README's and comments will follow!

WIP - DO NOT MERGE
JIRA: FUEL-207

Change-Id: I8e415aa8947b36e132f9578af882d975ef8b0abc
Signed-off-by: Jonas Bjurel <jonas.bjurel@ericsson.com>