Update patch set 1
Patch Set 1:

> Alexandru, I dont think this patch should go into stable/colorado.
 > There is less than 2 weeks before 2.0 is cut and it is to risky.
 > Colorado 2.0/3.0 is all about fixing critical bugs - not
 > improvements or feature adds.
 > Alex - look forward, this is part of the Danube release :-)

Hi, Jonas! Thank you for the feedback, I am ok with not merging this to stable.
I was just concerned that since the patch locations were changed, if we add patches to Fuel components on master,
we'll have to edit those by hand when cherry-picking to stable.
This happened the last time we changed the patch system, and a patch silently omitted on stable because the patch rework
was not cherry-picked beforehand.

If this has any chance of getting into C3, we'll take a fresh look at it then. If not, I will abandon it.

Patch-set: 1