Update patch set 1
Patch Set 1:

Hey Alex.
Stefan has a patch soon coming in on Master for Newton/Fuel10.
I think that this for the most will mean that any fixes we do for Colorado 2.0 and 3.0 will need to be pushed directly to stable/colorado - rather than cherry-picked from master. So in essence I don't see that we have much of an issue in having to tailor fixes for colorado, we will have to do that anyway going forward.
I don't think this will be in for colorado 3.0 either. Our focus should now be geared toward Danube and the mid November milestone which essentially means soft code-freeze for installers (I can be wrong on the milestone definition, but I think it is expected that we have stable Installer baselines from mid Nov).
BR/Jonas

Patch-set: 1