Update patch set 1
Patch Set 1:

This will pull in the head of master for fuel-plugin-ovs. I don't think this is what we want.

I think what we want is the head of stable/9.0 which is proposed by another Fuel review 23141.

The head of master has things like:

< export EXTRA_CFLAGS = -g -march=corei7-avx -mno-movbe -mno-abm -mno-fma -mno-bmi -mno-bmi2 -mno-avx2 -mno-lzcnt -mtune=corei7-avx

and 

<       sed -i '/RTE_MACHINE_CPUFLAG_AVX2/,+13 s/^/#/' lib/librte_acl/Makefile
<       sed -i '/__AVX2__/,+2 s/^/#/' mk/rte.cpuflags.mk

which leaves nearly all optimizations off - actually it conflicts with other lines which are also in master like:

 sed -ri 's,(RTE_MACHINE=).*,\1default,'

which will give a reasonable balance between performance and support as was agreed in comments on https://jira.opnfv.org/browse/FUEL-193

Patch-set: 1