Update patch set 5
Patch Set 5:

Hi,
My plan is to wait for upstream packetary to merge the patches described in the commit message here; then cherry-pick/adapt if necessary this to stable/colorado more or less in the current form (if you think this is something we want in Colorado, of course).

After that, I will patch this on master to completely drop fuel-mirror references, so it can be applied on top of Stefan's Newton rebase.

I tested building a bootstrap image (not only for amd64, but also for arm64) using this, and it seems to work nicely.
The size of the local mirror is slightly reduced (<800M instead of 1.1GB), the bootstrap build time is a little lower, but nothing significant.

Patch-set: 5
Label: Verified=0