Update patch set 6
Patch Set 6:

The idea here is to modularize the scenarios. We can dissect them to modules, e.g. in module-nfv-ovs.yaml, and then use them as building blocks for scenarios. Thus, less code duplication, because we can include 'module-nfv-osv.yaml' in every scenario we need it in. Does that make sense?

Patch-set: 6