fuel-library: pp: cobbler: Fix CentOS ver detect
Upstream fuel-library bug report: [1].

Recently, CentOS released version 7.3.1611, which triggered a latent
bug in fuel-library puppet scripts related to differentiating between
6.x and 7.x major versions: 7.3.1611 is wrongly detected as 6.x,
leading to the wrong packages being attempted to be installed,
leaving nailgun server service broken on the Fuel Master node.

This change forces the match at the beggining of the version string
(e.g. "/^6.+/: {" instead of "/6.+/: {"), so the new CentOS 7.3.1611
no longer matches the condition for 6.x).

Since we can't just bump the whole fuel-library git repo to the commit
including the fix (we still use Fuel 9.0.1, and not the tip of the
stable/colorado branches or at least Fuel 9.1), backport the fix as a
patch in Fuel@OPNFV.

The backported patch is safe to remove if/when we bump fuel-library
past upstream fix commit.

[1] https://bugs.launchpad.net/fuel/+bug/1649895

JIRA: FUEL-238

Change-Id: Id75c7085e02435a20ff4eeeedead3ad8a7010174
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>