Update patch set 1
Patch Set 1: Code-Review+1

Hi, Michael,
This is an AArch64 specific bit; I only pushed it because this particular patch causes constant rebase issues for the baremetal changes we are working on. It would simplify our work if it landed in Fuel@OPNFV.
No problem if you think it should stay in Armband for now.

Patch-set: 1
Label: Code-Review=+1