Update patch set 1
Patch Set 1: Verified-1

Not fully tested (requires baremetal deploy and we don't have x86 PODs, but I'll give it a try on AArch64 next week).
The motivation is we want to eventually support os-nosdn-nofeature-noha on baremetal, and we (Armband) would like to handle that scenario (only) without VCP, to accomodate machines with less resources.

Patch-set: 1
Label: Verified=-1