Update patch set 3
Patch Set 3: Code-Review+1

(5 comments)

This looks great!
There are some hardcodes left (eth0/eth1) in benchmark.yaml, which I'm not sure we use, and also in kvm.yaml, which define the interfaces that should be hooked to VCP VMs.
Later, we should probably move the interfaces names from arch-specific yaml to PDF, but that's a different story.

Patch-set: 3
Label: Code-Review=+1