Update patch set 1
Patch Set 1:

> > (1 comment)
 > 
 > I have not tried on a RHEL based Linux distro. Can't really say if
 > we are going to hit the same problem there

I just checked this on lf-pod2, `uuidgen` is provided by 'util-linux' package, so let's add that.
Also +1 to Michael's comment about commit message formatting. The recommended practice is to keep the title under 50 (or 52) columns, and the rest of the message wrapped at 70 (or 72) columns.

Patch-set: 1
Label: Workflow=0