ci/deploy.sh: Add PDF parsing dependencies
The following packages are required for PDF parsing:
- python
- python-ipaddress
- python-jinja2

Note that `python-ipaddress` is not available on older Ubuntu
distributions (e.g. Trusty), in which case `pip` should be used to
install it. An alternative would be to use python3 packages.

Update the installation instructions to specify minimum expected
requirements on the jump server.

JIRA: FUEL-292

Change-Id: I19fdcc163dcfdee45ac6fa3f5e09a28638d6035c
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>