deploy.sh: Check hypervisor after prereq install
On a clean system (i.e. no libvirt installed), the prerequisites
should be installed first, before checking for hypervisor access.

While at it, add 'libvirt' to the list of prerequisites, as it is
not automatically required by the others.

Add a note in the documentation about libvirtd service not being
automatically started after installation on CentOS.

Change-Id: I29d94d4a56cd823362811e8fc9c076c83fb6c04c
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>