[baremetal] MaaS: Enable HTTP transparent proxy
Instead of using NAT on the mas01 node for all cluster node outgoing
traffic, use the MaaS built-in transparent proxy so we can leverage
its caching capabilities too.
This allows us to drop the maas.pxe_nat state.

To use the new (non-transparent) MaaS proxy (:8000), add it for:
- default environment settings (via reclass);
- baremetal nodes salt minion config (via curtin template);
- TODO: VCP VMs salt minion config (via reclass, requires patching
  salt-formula-salt);

The MaaS proxy will not be used (at least for now) for:
- cfg01;
- mas01;

This change hardcodes MaaS-specific settings (the proxy) in the
common part of the reclass model, which will mostly likely require
a redesign when adding other baremetal provisioners (e.g. Ironic).

JIRA: FUEL-318

Change-Id: I6c5f6872bb94afb838580571080e808bc262fc68
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>