WIP Stacklight install
Some state changes will be made

[alav]
- add new deploy arg `-M` which enables Stacklight installation;
- use j2 templating for os-nosdn-nofeature-ha (only this one for now);
- adjust `do_templates` to support env vars during 1st stage parsing;
- TODO: rm os-nosdn-nofeature-stacklight-ha.yaml
- states: stacklight: use `notify`, minor style edits;
- move stacklight reclass dir to common-ha;
- add stacklight classes for ovs-ha scenario, based on
  common-ha/stacklight, move scenario specifics there;
- add stacklight state (conditional) for NOVCP scenarios;

NOTE: For NOVCP, a lot more changes are needed, including aligning
hostnames / IPs of StackLight node definitions with the existing
kvm (ctl) nodes.

Most likely, the current state of this change does NOT work and serves
only as a PoC for the desired integration.

JIRA: FUEL-308

Change-Id: I4bac4f460e8ebf834fcffbcc697d56b1b9e82a06
Signed-off-by: Delia Popescu <delia.popescu@enea.com>
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>