Update patch set 1
Patch Set 1: Workflow-1

(4 comments)

This change has at least 2 issues that need to be fixed.
I'm not sure how to continue with it - I can take care of the 2 issues when I find some more time - but I'm not confident I know enough Openstack to properly test this and see it through.

Patch-set: 1
Label: Workflow=-1