[NOVCP] Add deploy argument `-N` (experimental)
Instead of duplicating scenarios for NOVCP, allow it to be specified
using a new deploy argument, `-N`.

Things are getting convoluted, so instead of creating dedicated
'*_pdf.yml.j2' files for each group of similar features, apply the
templating in-place and rename all affected files to ".yml.j2".

This change only handles the templates for os-nosdn-nofeature-ha.

Change-Id: I7a80415ac33367ab227ececb4ffb1bc026546d36
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>