Update patch set 3
Patch Set 3:

> (2 comments)

The reason I did not add `ifconfig` was that I was worried about the security implications of making that public for the CI PODs.
However, all labs are behind proper firewalls, so I guess we can do it.
`lshw` requires the lshw package, and the information is hardly needed. The idea here is to dump info that helps us debug issues with certain deploys, not collect everything we can (we are not FB) ... lshw or other things won't help with DPDK information, as there's no guarantee the jumpserver resembles the baremetal nodes.

Patch-set: 3
Label: Verified=0