Update patch set 3
Patch Set 3:

> (3 comments)
 > 
 > More comments to follow. This patch is huge!

Ignore the whitespaces/missing copyright in that patch file, it will be upstreamed to Pharos and be gone from here.
As for vcpu count, you can have a VM with 1 VCPU, which is an odd number, so I don't see the issue :P We don't do guest cpu pinning to care about HT/NUMA there.

Patch-set: 3