Update patch set 1
Patch Set 1: Code-Review-1

This depends on upstream PR to be merged first.
Once that gets in, we can merge this, rebase Armband and drop one more Armband patch.
Note that this change will add arm64 repos to x86 nodes too when both architectures are used in the cluster, although this should not affect any x86 operations.

Patch-set: 1
Label: Code-Review=-1