Update patch set 1
Patch Set 1:

> The pinning still doesn't work for one package?

The pinning is very odd, it seems that maas is installed in the first batch of packages, so the pinning is not taken into account ...
I also tried installing maas formula from git, which works, but puts the _modules in a subdir, breaking our patching mechanism.
The simplest solution is to cherry-pick this change and hope that fabric-2 will always be on the 3rd interface; at least until a more recent baseline repo is created for salt formulas DEBs.

Patch-set: 1