Update patch set 1
Patch Set 1:

> > Submitted for discussion, not yet ready for merge.
 > > This is only useful when there are hybrid nodes and the virtual
 > > nodes get PXE IPs from MaaS DHCP.
 > 
 > but now we've static ip addresses ...

Yes, but those are applied by Salt once the minion registers with the master node, so we still use DHCP on PXE for bootstrapping the nodes first.
We could mess with user-data.template and inject the static IP before spawning the virtual nodes, haven't thought about that till now.

Patch-set: 1