[ha] AArch64: Add missing GPG key for armband repo
In a regular baremetal HA scenario, all cluster nodes are provisioned
via MaaS, which also adds the Enea GPG key to the APT keyring.
However, for a hybrid deployment, virtual nodes are not provisioned
by MaaS, so the GPG key should be added to reclass pillar data
explicitly for Salt to be able to import it.

JIRA: FUEL-338

Change-Id: I333b7c373bdbc631c0f54ac549f97ab889dec77c
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>