Update patch set 2
Patch Set 2: Workflow+1

> (1 comment)

No, this will only apply to virtual nodes when baremetal nodes are also present (otherwise the maas state file is not executed, so this new code block is not reached).
For baremetal deployments (where maas state is executed), the new `reset_vms` function will do nothing, as the only virtual nodes in that case are the infra VMs (cfg01, mas01).

Patch-set: 2
Label: Workflow=+1