Update patch set 1
Patch Set 1: Code-Review-1

The problem is with class inheritance in our reclass model and can be worked around in a simpler manner, instead of adding a new variable.
So, instead of setting the default value for barbican_integration to False and overriding it on the compute nodes to True, try the opposite.

Patch-set: 1
Label: Code-Review=-1