Update patch set 4
Patch Set 4:

> lgtm, waiting for another pair of eyes to review it
 > I would have a small nitpick on wrapping the commit message lines
 > at 70/72 columns, but I don't think it's worth a new revision just
 > for that

Hi Alexandru,

Sorry for the number of patches. I just realized that huge pages were not enabled on compute nodes. As a result I added the hugepages class as well. By this final change, I think that all requirements of EPA testcases are met.

I will run the installer with my patch and then run the testcases on that deployment to let you know if everything works as expected.

Regards,
Stamatis

Patch-set: 4