Update patch set 2
Patch Set 2:

> The corresponding patch has been merged into neutron formula.

We don't have neutron formula as submodule, so to bump it we'd need only to force a new docker build without the cached formulas.
Unfortunately, the 'best' solution I could find for this was to add a manual knob as [1].
I know this is not great, since it doesn't work like a pin, only a mininum baseline for the formulas, which might vary on container rebuild depending on existing local docker caches ...
All this so we won't run into non-applying patches due to upstream changes in formulas anymore.

[1] https://github.com/opnfv/fuel/blob/master/mcp/patches/docker/0002-OPNFV-package-installation-Ubuntu-user.patch#L36

Patch-set: 2