Update patch set 4
Patch Set 4:

This is not ready for a proper review, as there are a lot of FIXMEs left inside, but since this rewrite is quite large, I thought I should submit the latest draft now before polishing it a bit more tomorrow.

The overall structure should be close to the final form.
The developer guide (only imgs included in this change) will most likely not be ready before the docs milestone in 2 days, so we might leave it for the 7.1.0 release ...

Patch-set: 4