Update patch set 1
Patch Set 1:

> > > > > > I guess bgpvpn scenario isn't required/intended for
 > Gambia.
 > > > > >
 > > > > > That is correct
 > > > >
 > > > > Actually, fuel support is among the goals of SDNVPN for
 > Gambia
 > > > > release[1].
 > > > > Do you see any obstacle on this?
 > > > >
 > > > > [1] https://wiki.opnfv.org/display/sdnvpn/OPNFV+SDNVPN+Gambia+Release+Plan
 > > >
 > > > This scenario should have been added in advance to release it
 > > > officially. Anyway I don't mind to merge it now since there is
 > no
 > > > some impact for other functionally.
 > >
 > > I agree, let's merge it and see later if we can apply for a
 > > milestone exception for later point releases (7.1/7.2) to make it
 > > official.
 > 
 > Actually I would prefer to ask for the exception before merging

Hi there, I have a question which is either relevant or irrelevant (I am not sure): Supposing that the scenario is merged in stable/gambia, then a new jenkins job should be created to run it like all the other scenarios? Or this is not related with the presence or not of bgpvpn scenario in fuel's gambia branch?

Patch-set: 1
Reviewer: Gerrit User 7381 <7381@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=0