Update patch set 51
Patch Set 51:

> lgtm. We'll also need to add some docs for this new scenarios.
 > About the failing verify - don't mind that, one of the changes we
 > tested recently messed with the owner/permissions of some files in
 > the Jenkins workspace dir, so subsequent runs had issues removing
 > those.

Hi Alex, thanks for the review. According to the documentation - we are planning to describe installation process in Auto documentation in detail. I suppose, that in case of Fuel documentation, it will be enough to introduce appropriate directories for new scenarios into docs/release/scenarios/ with some brief description (following the structure of existing scenarios) with link to the related chapter of Auto documentation. We will prepare separate patches for documentation in next weeks. I hope this would be sufficient.

Patch-set: 51