Update patch set 2
Patch Set 2:

> It works so far and start failing? why?
 > Barbican doesn't depends on order of nodes to install.

It only fails on arm-pod10, which is a mixed-arch POD with x86 kvm nodes - not really sure why though. The behavior is consistent, i.e. if the barbican state runs on ctl02/ctl03, it will fail with a missing table column in the DB ... but once it runs on ctl01, everything works fine.

Patch-set: 2