Update patch set 1
Patch Set 1:

So for Gambia the problem seems to be with OVS-DPDK and not ODL (as ODL without DPDK is fine).
I did a quick local test and I was able to reproduce the issue on one of the computes nodes - the dpdk0 port fails to bind with:
  root@cmp002:~# ovs-vsctl show
    [...]
        Port "dpdk0"
            Interface "dpdk0"
                type: dpdk
                options: {dpdk-devargs="0000:00:05.0", n_rxq="2"}
                error: "could not add network device dpdk0 to ofproto (No such device)"
This corresponds to a dmesg crash:
  [ 2295.796221] pmd7[6511]: segfault at 0 ip 0000000000545a99 sp 00007fcfa13f9860 error 4 in ovs-vswitchd-dpdk[400000+22d000]

No idea what is triggering it though and why it didn't happen before. Restarting openvswitch-server binds the ports correctly afaict.

Patch-set: 1
Label: Code-Review=0