Update patch set 1
Patch Set 1:

Hi, Parker,
I didn't test this enough, and it's not exactly a fix for the rootcause in UNH lab (I can't pinpoint what's happening there, as everything looks good in the logs and works just fine when re-run manually); BUT it should be a decent workaround.
When you have some time, it'd be great if we could apply this patch locally and try a clean deploy with it, to confirm it actually works around the problem.

Patch-set: 1
Reviewer: Gerrit User 2997 <2997@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0