Update patch set 7
Patch Set 7:

> After each verify/deploy the others CI jobs getting failures due to
 > "unable to remove docker network".

Yes, previously 10.20.0.0/24 was assigned to a virsh-managed network, and this patch changes it to a Docker network. On subsequent deploys,  we try to create the virsh networks before cleaning up the Docker networks ...
If we merge this, it should stop happening, it only breaks when deploying older versions after a deploy was done with this one.

Patch-set: 7
Reviewer: Gerrit User 2997 <2997@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0